+++ CHANGES.rst
@@ -20,6 +20,8 @@ Bug fixes:
 
 - Fix use of add_file in testbrowser tests. [davisagli]
 
+- Render migration results without using Zope session. [davisagli]
+
 
 1.2.20 (2017-01-20)
 -------------------